convert lower motor mount for bearings that replace pulley flange
there is no good way to get the belt as close as desired to the 2040
tower with both a sufficient double shear support bearing and the 1.5
mm flange on the pulley. the current build is using a 5x10x3 bearing,
which is undersized to begin with, and only works with a shaved-down
pulley or a longer motor shaft.

instead, take off the pulley flange to get 1.5 mm of engagement in a
bearing with 8 mm ID. a flanged bearing, mf148, is used so that we can
get the bearing as close as possible to the 2040 without it scraping.
otherwise, the plastic would need to provide the spacing, which would
require a separate insert part to get the right orientation to be that
thin.

the limited engagement length is potentially problematic, but better
than the current situation, and can be augmented with a (printed or
metal) spacer/bushing (5 mm bore, 8 mm OD).

the hypothetical 9 mm belt version instead uses mf128 to get the belt
closer to the extrusion. this may be changed if a real 9 mm build is
attempted at some point. depending on how coupling to the carriage
works out for 9 mm, it may be possible to use the mf148 there too.