fix belt linearity at carriage attachment
the old value 3.15 was probably either guessed, measured poorly, or
wrongly figured from the back surface of the belt rather than the
pitch line. update it to match the actual pitch line from the pulleys.
this corrects a linearity error of 0.5 mm.

this may still need further tuning to compensate for the fact that the
belt where it enters the carriage is mated with its own opposing teeth
rather than a pulley, which will not necessarily respect the same
pitch line.

as a separate issue, it currently does not match the belt mockup,
because the mockup is based on gt2 dimensions, not the thicker gt3
belt the carriage is currently configured for. however, even with that
corrected for, they don't entirely match; this is probably because the
carriage CAD is using the tooth height without PLD incorporated to
figure the pitch line.