begin factoring parameters out of delta vertex
originally the vertices were constructed by CSG subtracting out a
full, 3-tower negative of the extrusions and screws from a base block.
this is not only inefficient, but gratuitously carried a dependence on
the frame size parameters (which cancelled out with themselves)
despite the vertices being intended to be frame-size-invariant parts.

this commit factors the cancelling XY translations out of the column
negative, but still leaves the height dependency and everything for
the end extrusions in place.